mv bowtie,hisat,fastqc fixtures to respective test modules
