+++ wrappers/test/test_bowtie2.py
@@ -1,32 +1,34 @@
+import os
 import pytest
-import pysam
 from snakemake.shell import shell
 from lcdblib.snakemake import aligners
-from utils import run, dpath, rm, symlink_in_tempdir
+from utils import run, dpath, symlink_in_tempdir, tmpdir_for_func
 
 
-def test_bowtie2_build(dm6_fa, tmpdir):
+@pytest.fixture(scope='session')
+def bowtie2_indexes(dm6_fa, tmpdir_factory):
+    d = tmpdir_for_func(tmpdir_factory)
     snakefile = '''
-                rule bowtie2_build:
-                    input:
-                        fasta='2L.fa'
-                    output:
-                        index=expand('data/assembly/assembly.{n}.bt2', n=range(1,5))
-                    log: 'bowtie2.log'
-                    wrapper: "file:wrapper"
-
-                '''
-    input_data_func=symlink_in_tempdir(
+    rule bowtie2:
+        input: fasta='dm6.fa'
+        output: index=['dm6.1.bt2', 'dm6.2.bt2']
+        log: 'bowtie2.log'
+        wrapper: 'file:wrapper'
+    '''
+    input_data_func = symlink_in_tempdir(
         {
-            dm6_fa: '2L.fa'
+            dm6_fa: 'dm6.fa'
         }
     )
 
     def check():
         assert 'Total time for backward call to driver' in open('bowtie2.log').readlines()[-1]
-        assert list(shell('bowtie2-inspect data/assembly/assembly -n', iterable=True)) == ['2L']
+        assert list(shell('bowtie2-inspect dm6 -n', iterable=True)) == ['2L', '2R']
 
-    run(dpath('../wrappers/bowtie2/build'), snakefile, check, input_data_func, tmpdir)
+    run(
+        dpath('../wrappers/bowtie2/build'),
+        snakefile, check, input_data_func, d)
+    return aligners.bowtie2_index_from_prefix(os.path.join(d, 'dm6'))
 
 
 def _dict_of_bowtie2_indexes(bowtie2_indexes, prefix):
